ai-server/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
PyPDF2==3.0.1
langchain==0.2.14
langchain-google-genai==1.0.6
langchain-community==0.2.12
faiss-cpu==1.8.0
google-generativeai==0.7.2
python-dotenv==1.0.1
